services/summary_service/requirements.txt
aiokafka==0.10.0
python-dotenv==1.0.1
pydantic==1.10.15
requests==2.32.5

